fix error

+++ src/DataReader.java
@@ -2,7 +2,7 @@ import java.util.ArrayList;
 
 public class DataReader extends TextFileReader {
     private static double[][] nodes= new double[16][2];
-    private static double[] demands= new double [16];
+    private static int[] demands= new int [16];
     private static double[][] truckLoad=  new double[4][2];
     private static double[] productWeight= new double[5];
     private static TextFileReader nodeFile;
@@ -33,7 +33,7 @@ public class DataReader extends TextFileReader {
     {
         return nodes;
     }
-    public static double[] getDemands()
+    public static int[] getDemands()
     {
         return demands;
     }
